+++ src/or/directory.c
@@ -85,7 +85,7 @@ static int directory_send_command(connection_t *conn, int command) {
         return -1;
       }
       snprintf(tmp, sizeof(tmp), "POST / HTTP/1.0\r\nContent-Length: %d\r\n\r\n%s",
-               strlen(s), s);
+               (int)strlen(s), s);
       connection_write_to_buf(tmp, strlen(tmp), conn);
       conn->state = DIR_CONN_STATE_CLIENT_SENDING_UPLOAD;
       break;
